Improve flag handling, cleanup UI trait
Add n and l to editor state, toggling if numbering/literal is default or not.
Add commands 'N' and 'L' as well as flags to Ed::new to configure these.
Rename UI::print to print_message, to clarify its intended use.
Remove EdState argument from UI::print_message, since unlikely to be used.
Fix 's' so that printing flags are stored with last invocation
Fix 's' so that error is returned if no match is found
Refactor some code and comments.