Change return type of Buffer.get_selection
Changing it to a concrete type, rather than an unspecified Iterator, allows for
creating a dyn IO (since a vtable cannot be constructed over a trait unless the
input types are concrete). Could have made it into a dyn Iterator, but calling
via a vtable for every line in a file would probably have a noticeable impact on
performance. So in the end we are now exporting an ugly, implementation
specific type in our public API. Not great, but that implementation _should_ be
fine for a while at least.