Add text validation to Line and PubLine via new LineText struct
And the refactor this necessitated.
This was needed so that we can trust our internal strings to always be newline
terminated, no matter the input we are given.

Arguably it would be better to handle both cases, but at some point we will have
to draw the line so I'm doing it here now. If someone wishes to add tests for
input with and without terminating newlines and patch up to make them pass I
won't stop them, but I won't do it while we still need to verify that there are
no newlines within the lines.

Also squeezed in to encapsulate the tag field of Line, to prevent misuse and
make using it simpler and more concise.