feat: Improve index handling
Commands that take a lone index, either as selection or argument, now handle
these a bit better. Correctness is up for debate, as always, but I'd argue that
this behaviour is slightly more useful (append 2 lines after current selection
seems better than append 2 lines after the start of the current selection to me)

Flagged as a feature since it changes behaviour of the affected commands.