feat!: Add `:` index
It selects the other index in selection, so the starting index if used in the
ending position and the ending index if used in the starting position.
This also can be used to use the non-default index for commands that only take
a singular index, for example `:a` appends after the first line in current
selection, or which take index arguments, `,3m:` moves to after first line in
current selection.